Move CanvasContextSpy deeper alongside the tests that use it.
