+++ test/layers/CanvasContextSpy-test.ts
