+++ test/layers/CanvasContextSpy.ts
